Fixed TeleopAlignWithTape
So I've moved the value for the pid setPoints for the centerX and area  to the visionSystem. I also changed the alignWithTape command to run while the button is pressed. Whenever the button is let go, the command is canceled. I also added in a boolean that cancels the command if both the x and area reach an acceptable threshold. The robot aligns with the vision tape as it did before we moved the pipeline into the project. Thank you for helping us restructure the project!