Climb sequence *mostly* works.
Needs some tuning. Keanu needs to read these commits and work on tuning this code this week. Still need to bring the climber back up, as well as tune PID for the climber before it starts driving (oscillations due to PID). I have a video that I will send to George of the climb sequence working.